labs/extra/byodapp/requirements.txt
fastapi
uvicorn
python-dotenv
tiktoken
openai
langchain
qdrant-client